fix: improve RECREATE macro
This macro was actually fine the way it was because it aborts if the realloc call fails.
However, if the macro were ever changed to handle the error instead of aborting,
there would be a potential memory leak in that the original pointer would be lost.
Fixing it here is mostly because it was causing a clang-tidy warning that I found
annoying, but it also future proofs it if it's refactored to handle the error.